requirements.txt
openai
pandas
openpyxl
pillow